+++ R/count.R
@@ -117,11 +117,10 @@ step_count_new <-
 #' @export
 prep.step_count <- function(x, training, info = NULL, ...) {
   col_name <- recipes_eval_select(x$terms, training, info)
+  check_type(training[, col_name], quant = FALSE)
+
   if (length(col_name) > 1) {
     rlang::abort("The selector should select at most a single variable")
-  }
-  if (any(info$type[info$variable %in% col_name] != "nominal")) {
-    rlang::abort("The regular expression input should be character or factor")
   }
 
   step_count_new(