+++ R/holiday.R
@@ -91,11 +91,11 @@ prep.step_holiday <- function(x, training, info = NULL, ...) {
   col_names <- recipes_eval_select(x$terms, training, info)
 
   holiday_data <- info[info$variable %in% col_names, ]
-  if (any(holiday_data$type != "date")) {
+  if (any(!holiday_data$type %in% c("date", "datetime"))) {
     rlang::abort(
       paste0(
-        "All variables for `step_holiday` should be either `Date` ",
-        "or `POSIXct` classes."
+        "All variables for `step_holiday` should be either `Date` or",
+        "`POSIXct` classes."
       )
     )
   }