+++ R/unknown.R
@@ -87,17 +87,7 @@ step_unknown_new <-
 #' @export
 prep.step_unknown <- function(x, training, info = NULL, ...) {
   col_names <- recipes_eval_select(x$terms, training, info)
-  col_check <- dplyr::filter(info, variable %in% col_names)
-  if (any(col_check$type != "nominal")) {
-    rlang::abort(
-      paste0(
-        "Columns must be character or factor: ",
-        paste0(col_check$variable[col_check$type != "nominal"],
-          collapse = ", "
-        )
-      )
-    )
-  }
+  check_type(training[, col_names], quant = FALSE)
 
   # Get existing levels and their factor type (i.e. ordered)
   objects <- lapply(training[, col_names], get_existing_values)