+++ R/window.R
@@ -187,10 +187,7 @@ step_window_new <-
 #' @export
 prep.step_window <- function(x, training, info = NULL, ...) {
   col_names <- recipes_eval_select(x$terms, training, info)
-
-  if (any(info$type[info$variable %in% col_names] != "numeric")) {
-    rlang::abort("The selected variables should be numeric")
-  }
+  check_type(training[, col_names])
 
   if (!is.null(x$names)) {
     if (length(x$names) != length(col_names)) {