+++ tests/testthat/test_roles.R
@@ -9,7 +9,10 @@ test_that("default method", {
   rec <- recipe(x = biomass)
   exp_res <- tibble(
     variable = colnames(biomass),
-    type = rep(c("nominal", "numeric"), c(2, 6)),
+    type = rep(
+      list(c("string", "unordered", "nominal"), c("double", "numeric")),
+      c(2, 6)
+    ),
     role = NA_character_,
     source = "original",
     required_to_bake = TRUE
@@ -22,7 +25,10 @@ test_that("changing roles", {
   rec <- update_role(rec, sample, new_role = "some other role")
   exp_res <- tibble(
     variable = colnames(biomass),
-    type = rep(c("nominal", "numeric"), c(2, 6)),
+    type = rep(
+      list(c("string", "unordered", "nominal"), c("double", "numeric")),
+      c(2, 6)
+    ),
     role = rep(c("some other role", NA), c(1, 7)),
     source = "original",
     required_to_bake = TRUE
@@ -42,7 +48,10 @@ test_that("change existing role", {
 
   exp_res <- tibble(
     variable = colnames(biomass),
-    type = rep(c("nominal", "numeric"), c(2, 6)),
+    type = rep(
+      list(c("string", "unordered", "nominal"), c("double", "numeric")),
+      c(2, 6)
+    ),
     role = rep(c("other other role", NA), c(1, 7)),
     source = "original",
     required_to_bake = TRUE
@@ -99,7 +108,10 @@ test_that("new role for existing NA role", {
 
   exp_res <- tibble(
     variable = colnames(biomass),
-    type = rep(c("nominal", "numeric"), c(2, 6)),
+    type = rep(
+      list(c("string", "unordered", "nominal"), c("double", "numeric")),
+      c(2, 6)
+    ),
     role = rep(c("some other role", NA), c(1, length(colnames(biomass)) - 1)),
     source = "original",
     required_to_bake = TRUE
@@ -114,7 +126,12 @@ test_that("new role with specified type", {
 
   exp_res <- tibble(
     variable = c("sample", colnames(biomass)),
-    type = c("nominal", "new type", "nominal", rep("numeric", 6)),
+    type = c(
+      list(c("string", "unordered", "nominal")),
+      list("new type"),
+      list(c("string", "unordered", "nominal")),
+      rep(list(c("double", "numeric")), 6)
+    ),
     role = rep(c("blah", "some other role", NA), c(1, 1, 7)),
     source = "original",
     required_to_bake = TRUE
@@ -132,7 +149,13 @@ test_that("add new role when two already exist with different types", {
 
   exp_res <- tibble(
     variable = c("sample", "sample", colnames(biomass)),
-    type = c("nominal", "new type", "nominal", "nominal", rep("numeric", 6)),
+    type = c(
+      list(c("string", "unordered", "nominal")),
+      list("new type"),
+      list(c("string", "unordered", "nominal")),
+      list(c("string", "unordered", "nominal")),
+      rep(list(c("double", "numeric")), 6)
+    ),
     role = c("blah", "some other role", "another role", rep(NA, 7)),
     source = "original",
     required_to_bake = TRUE
@@ -153,7 +176,10 @@ test_that("existing role is skipped", {
   # also tests the order, new roles come directly after old ones
   exp_res <- tibble(
     variable = c("sample", colnames(biomass)),
-    type = rep(c("nominal", "numeric"), c(3, 6)),
+    type = rep(
+      list(c("string", "unordered", "nominal"), c("double", "numeric")),
+      c(3, 6)
+    ),
     role = rep(c("blah", "some other role", NA), c(1, 1, 7)),
     source = "original",
     required_to_bake = TRUE
@@ -176,7 +202,10 @@ test_that("existing role is skipped, but new one is added", {
       rep(c("sample", "dataset"), c(2, 2)),
       setdiff(colnames(biomass), c("sample", "dataset"))
     ),
-    type = rep(c("nominal", "numeric"), c(4, 6)),
+    type = rep(
+      list(c("string", "unordered", "nominal"), c("double", "numeric")),
+      c(4, 6)
+    ),
     role = c("blah", "some other role", NA, "some other role", rep(NA, 6)),
     source = "original",
     required_to_bake = TRUE
@@ -230,7 +259,10 @@ test_that("remove roles", {
 
   exp_res <- tibble(
     variable = colnames(biomass),
-    type = rep(c("nominal", "numeric"), c(2, 6)),
+    type = rep(
+      list(c("string", "unordered", "nominal"), c("double", "numeric")),
+      c(2, 6)
+    ),
     role = NA_character_,
     source = "original",
     required_to_bake = TRUE
@@ -245,7 +277,12 @@ test_that("New type for an existing role can be added", {
 
   exp_res <- tibble(
     variable = c("sample", colnames(biomass)),
-    type = c(c("nominal", "text", "nominal"), rep("numeric", 6)),
+    type = c(
+      list(c("string", "unordered", "nominal")),
+      list("text"),
+      list(c("string", "unordered", "nominal")),
+      rep(list(c("double", "numeric")), 6)
+    ),
     role = c("role1", "role1", rep(NA, 7)),
     source = "original",
     required_to_bake = TRUE
@@ -431,7 +468,7 @@ test_that("Existing `NA` roles are not modified in prep() when new columns are g
     filter(new, grepl("Species", variable)),
     tibble(
       variable = c("Species_versicolor", "Species_virginica"),
-      type = rep("numeric", times = 2),
+      type = rep(list(c("integer", "numeric")), 2),
       role = rep("predictor", times = 2),
       source = rep("derived", times = 2)
     )